v.1.3.4 review_1

+++ src/main/java/ru/yandex/practicum/filmorate/model/Like.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class Like {
-    private long userId;
-    private long filmId;
+    private int userId;
+    private int filmId;
 }
